+++ types/global.d.ts
@@ -8,7 +8,7 @@ declare function App<G>(options: tinyapp.AppOptions<G>): void;
 /**
  * 获取小程序实例，一般用在各个子页面之中获取顶层应用。
  */
-declare function getApp(): { globalData: any };
+declare function getApp(): tinyapp.IGetAppResult;
 
 /**
  * Page() 函数用来注册一个页面。